pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asciinema>=2.4.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.0",
    "pyyaml>=6.0.2",
    "ruff>=0.11.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gendiff"]

[project.scripts]
gendiff = "gendiff.scripts.gendiff:main"
